+++ src/utils/signal_handler.c
@@ -6,11 +6,12 @@
 /*   By: emadriga <emadriga@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/01/05 19:58:37 by emadriga          #+#    #+#             */
-/*   Updated: 2022/01/31 09:05:36 by emadriga         ###   ########.fr       */
+/*   Updated: 2022/02/08 21:58:22 by emadriga         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "minishell.h"
+#define SIGQUIT_MSG "Quit:"
 
 /**
  * * Handles SIGINT signal CTRL+C, promping new Line
@@ -19,7 +20,8 @@
 static void	signal_handler_default_sigint(int signal)
 {
 	(void)signal;
-	write(STDOUT_FILENO, "\n", 1);
+	g_var.last_status = 130;
+	ft_putstr_fd("\n", STDOUT_FILENO);
 	rl_replace_line("", 0);
 	rl_on_new_line();
 	rl_redisplay();
@@ -31,16 +33,10 @@ static void	signal_handler_default_sigint(int signal)
 */
 void	signal_handler_forks(int is_child)
 {
-	signal(SIGINT, SIG_DFL);
 	if (is_child == TRUE)
-	{
-		signal(SIGQUIT, SIG_DFL);
-	}
+		signal(SIGINT, SIG_DFL);
 	else
-	{
 		signal(SIGINT, SIG_IGN);
-		signal(SIGQUIT, SIG_IGN);
-	}
 }
 
 /**
@@ -67,10 +63,17 @@ void	signal_handler_elephants_sigint(int signal)
  * * Handles SIGINT signal CTRL+C, exit process
  * @param signal	signal identifier (see man signal)
 */
-void signal_handler_process_sigint(int signal)
+void	signal_handler_process_sigint(int signal)
 {
-	(void)signal;
-	g_var.current_status = 130;
-	exit(g_var.current_status);
+	if (signal == SIGINT)
+	{
+		g_var.current_status = 130;
+		ft_putstr_fd("\n", STDOUT_FILENO);
+	}
+	else if (signal == SIGQUIT)
+	{
+		g_var.current_status = 131;
+		ft_putstr_fd(SIGQUIT_MSG, STDOUT_FILENO);
+		ft_putstr_fd("\n", STDOUT_FILENO);
+	}
 }
-