+++ CHANGELOG.md
@@ -5,6 +5,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+
+## 2022-02-08 - Edu
+### Added
+- Handled SIGQUIT on processes
+
+### Changed
+- Reduce complexity run_processes
+
 ## 2022-01-31 - Edu
 ### Changed
 - Modify ft-rand and renamed to ft_rand_unit for our purposes